algunos errores arreglados

+++ app/Http/Controllers/UsuarioController.php
@@ -15,7 +15,7 @@ class UsuarioController extends Controller
     {
         // with('rol') carga la relación y evita el problema de consultas N+1
         $usuarios = Usuario::with('rol')->orderBy('nombre')->get();
-        return view('usuarios.index', compact('usuarios'));
+        return view('adminpanel.usuarios.index', compact('usuarios'));
     }
 
     /**
@@ -24,7 +24,7 @@ class UsuarioController extends Controller
     public function create()
     {
         $roles = Rol::all(); // necesario para el <select> del formulario
-        return view('usuarios.create', compact('roles'));
+        return view('adminpanel.usuarios.create', compact('roles'));
     }
 
     /**
@@ -48,8 +48,13 @@ class UsuarioController extends Controller
      * Remove the specified resource from storage.
      */
     public function destroy(Usuario $usuario)
-    {
-        $usuario->delete(); // borrado lógico
-        return redirect()->route('usuarios.index')->with('exito', 'Usuario dado de baja.');
+{
+    // Evita que el usuario autenticado se borre a sí mismo
+    if ($usuario->id === auth()->id()) {
+        return redirect()->route('usuarios.index')->with('error', 'Acción denegada: No podés dar de baja una cuenta de Administrador');
+    }
+
+    $usuario->delete(); // borrado lógico
+    return redirect()->route('usuarios.index')->with('exito', 'Usuario dado de baja.');
     }
 }
